Refactor into separate files
